+++ packages/akashic-cli-export/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @akashic/akashic-cli-export
 
+## 2.0.5
+
+### Patch Changes
+
+- [#1553](https://github.com/akashic-games/akashic-cli/pull/1553) [`1165411`](https://github.com/akashic-games/akashic-cli/commit/1165411463a9567bc853a2e0c93b4a69f6f1ae62) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@akashic/headless-driver` to `2.17.4`.
+
+## 2.0.4
+
+### Patch Changes
+
+- [#1483](https://github.com/akashic-games/akashic-cli/pull/1483) [`c0fdad7`](https://github.com/akashic-games/akashic-cli/commit/c0fdad7105d41ebf6ce8d9b1ffab13f91df54260) Thanks [@renovate](https://github.com/apps/renovate)! - Updated dependency `@akashic/headless-driver` to `2.17.3`.
+
+## 2.0.3
+
+### Patch Changes
+
+- [#1546](https://github.com/akashic-games/akashic-cli/pull/1546) [`50e76c1`](https://github.com/akashic-games/akashic-cli/commit/50e76c18698d40f65aca5c422f0e1732b42dc8b6) Thanks [@ShinobuTakahashi](https://github.com/ShinobuTakahashi)! - Remove imageAsset hint.untainted in export-html
+
 ## 2.0.2
 
 ### Patch Changes